requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
moviepy==1.0.3
Pillow==10.4.0
tensorflow==2.15.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
decorator==4.4.2
tqdm==4.66.4
requests==2.32.3
proglog==0.1.10